reorg for readability, cleaned out init, added config
