+++ changefeedtrigger/function.json
@@ -5,7 +5,7 @@
           "name": "mytimer",
           "type": "timerTrigger",
           "direction": "in",
-          "schedule": "0 * * * *"
+          "schedule": "*/1 * * * *"
       }
   ]
 }
