+++ changefeedtrigger/requirements.txt
@@ -1,0 +1,2 @@
+azure-functions==1.3.0
+azure-storage-blob-changefeed==12.0.0b1
